resolvconf -> /sbin/resolvconf and a new reset feature
These changes do two things. One is a minor fix (resolvconf to /sbin/resolvconf) for using this under sudo instead of su. The other is a new feature that lets you call the file as a script to reset the firewall. I needed this because I don't use the vpn full-time. I've also added documentation to make sure the difference between the original use as a config script and the new use as a user script is clear.